# added cached tokenizer