@ArthurZucker 2024-12-23
Also you could compute with both slow and fast tokenizers, I don't know which you are using!