@Deep-unlearning 2025-01-16
> Also you could compute with both slow and fast tokenizers, I don't know which you are using!